[NCL-544] Merge model builder classes into main model classes
Reduce some boilerplate code and follow the common builder pattern
